requirements.txt
streamlit
pinecone
groq
sentence-transformers
python-dotenv
requests
